refactor(frontend): extract lookup api module
